refactor to reduce future trust updates
