+++ JamfUploaderProcessors/JamfIconUploader.py
@@ -1,27 +1,26 @@
 #!/usr/local/autopkg/python
 
 """
-JamfIconUploader processor for uploading a category to Jamf Pro using AutoPkg
+JamfIconUploader processor for uploading an icon to Jamf Pro using AutoPkg
     by G Pugh
 """
 
 import os.path
 import sys
 
-from time import sleep
-from autopkglib import ProcessorError
-
 # to use a base module in AutoPkg we need to add this path to the sys.path.
 # this violates flake8 E402 (PEP8 imports) but is unavoidable, so the following
 # imports require noqa comments for E402
 sys.path.insert(0, os.path.dirname(__file__))
 
-from JamfUploaderLib.JamfUploaderBase import JamfUploaderBase  # noqa: E402
+from JamfUploaderLib.JamfIconUploaderBase import (  # noqa: E402
+    JamfIconUploaderBase,
+)
 
 __all__ = ["JamfIconUploader"]
 
 
-class JamfIconUploader(JamfUploaderBase):
+class JamfIconUploader(JamfIconUploaderBase):
     """A processor for AutoPkg that will upload an icon to a Jamf Cloud or on-prem server.
     Note that an icon can only be successsfully injected into a Mac App Store app item if
     Cloud Services Connection is enabled."""
@@ -80,130 +79,12 @@ class JamfIconUploader(JamfUploaderBase):
         },
     }
 
-    def get_icon(self, icon_uri):
-        """download an icon file"""
-
-        self.output(f"Downloading icon from {icon_uri}...", verbose_level=2)
-        # download the icon
-        count = 0
-        while True:
-            count += 1
-            self.output(
-                f"Icon download attempt {count}",
-                verbose_level=2,
-            )
-            request = "GET"
-            r = self.curl(request=request, url=icon_uri, endpoint_type="icon_get")
-            # check HTTP response
-            if self.status_check(r, "Icon", icon_uri, request) == "break":
-                break
-            if count > 5:
-                self.output("ERROR: Icon download did not succeed after 5 attempts")
-                self.output(f"\nHTTP POST Response Code: {r.status_code}")
-                raise ProcessorError("ERROR: Icon download failed ")
-            if int(self.sleep) > 30:
-                sleep(int(self.sleep))
-            else:
-                sleep(30)
-        return r
-
-    def upload_icon(self, jamf_url, icon_file, token):
-        """Upload icon."""
-
-        self.output("Uploading icon...")
-
-        # if we find an object ID we put, if not, we post
-        object_type = "icon"
-        url = "{}/{}".format(jamf_url, self.api_endpoints(object_type))
-
-        # upload the icon
-        count = 0
-        while True:
-            count += 1
-            self.output(
-                f"Icon upload attempt {count}",
-                verbose_level=2,
-            )
-            request = "POST"
-            r = self.curl(
-                request=request,
-                url=url,
-                token=token,
-                data=icon_file,
-                endpoint_type="icon_upload",
-            )
-
-            # check HTTP response
-            if self.status_check(r, "Icon", icon_file, request) == "break":
-                break
-            if count > 5:
-                self.output("ERROR: Icon upload did not succeed after 5 attempts")
-                self.output(f"\nHTTP POST Response Code: {r.status_code}")
-                raise ProcessorError("ERROR: Icon upload failed ")
-            if int(self.sleep) > 30:
-                sleep(int(self.sleep))
-            else:
-                sleep(30)
-        return r
+    description = __doc__
 
     def main(self):
-        """Do the main thing here"""
-        self.jamf_url = self.env.get("JSS_URL")
-        self.jamf_user = self.env.get("API_USERNAME")
-        self.jamf_password = self.env.get("API_PASSWORD")
-        self.client_id = self.env.get("CLIENT_ID")
-        self.client_secret = self.env.get("CLIENT_SECRET")
-        self.icon_file = self.env.get("icon_file")
-        self.icon_uri = self.env.get("icon_uri")
-        self.sleep = self.env.get("sleep")
+        """Run the execute function"""
 
-        # clear any pre-existing summary result
-        if "jamficonuploader_summary_result" in self.env:
-            del self.env["jamficonuploader_summary_result"]
-
-        # get token using oauth or basic auth depending on the credentials given
-        if self.jamf_url and self.client_id and self.client_secret:
-            token = self.handle_oauth(self.jamf_url, self.client_id, self.client_secret)
-        elif self.jamf_url and self.jamf_user and self.jamf_password:
-            token = self.handle_api_auth(
-                self.jamf_url, self.jamf_user, self.jamf_password
-            )
-        else:
-            raise ProcessorError("ERROR: Credentials not supplied")
-
-        # obtain the icon from the URI if no file path provided
-        if (
-            "https://ics.services.jamfcloud.com/icon" in self.icon_uri
-            and not self.icon_file
-        ):
-            r = self.get_icon(self.icon_uri)
-            self.icon_file = r.output
-
-        if not self.icon_file:
-            raise ProcessorError("ERROR: Icon not found")
-
-        # upload the icon
-        r = self.upload_icon(
-            self.jamf_url,
-            self.icon_file,
-            token,
-        )
-
-        # get the uri from the output
-        self.selfservice_icon_uri = r.output["url"]
-        self.icon_id = r.output["id"]
-
-        # output the summary
-        self.env["selfservice_icon_uri"] = self.selfservice_icon_uri
-        self.env["icon_id"] = str(self.icon_id)
-        self.env["jamficonuploader_summary_result"] = {
-            "summary_text": "The following icons were uploaded in Jamf Pro:",
-            "report_fields": ["selfservice_icon_uri", "icon_id"],
-            "data": {
-                "selfservice_icon_uri": self.selfservice_icon_uri,
-                "icon_id": str(self.icon_id),
-            },
-        }
+        self.execute()
 
 
 if __name__ == "__main__":
